Add basic testing skeleton for new library. Still needs mocking.
