+++ test/testdata/ENV.json
@@ -1,0 +1,9 @@
+{
+    "PLATFORM_APP_DIR": "/app",
+    "PLATFORM_APPLICATION_NAME": "app",
+    "PLATFORM_PROJECT": "test-project",
+    "PLATFORM_TREE_ID": "abc123",
+    "PLATFORM_PROJECT_ENTROPY": "def789",
+
+    "SOME_VARIABLE": "some value"
+}
