+++ test/testdata/PLATFORM_VARIABLES.json
@@ -1,0 +1,3 @@
+{
+    "somevar": "someval"
+}
